Осталось 4 ошибки Postman

+++ backend/recipes/views.py
@@ -2,6 +2,7 @@ import os
 
 from django.contrib.auth import get_user_model
 from django.shortcuts import get_object_or_404
+from django_filters.rest_framework import DjangoFilterBackend
 from dotenv import find_dotenv, load_dotenv
 from hashids import Hashids
 from recipes.pagination import SubscribePagination, RecipePagination
@@ -15,7 +16,7 @@ from .serializers import (FavoriteSerializer, IngredientSerializer,
                           RecipeGetSerializer, RecipePostSerializer,
                           ShoppingCartSerializer, SubscribeSerializer,
                           TagSerializer)
-
+from recipes.filters import IngredientFilter, RecipeFilter
 User = get_user_model()
 load_dotenv(find_dotenv())
 SERVER_HOST = os.getenv("SERVER_DOMEN")
@@ -35,10 +36,22 @@ class SubscribeViewSet(viewsets.ModelViewSet):
         author_id = self.kwargs['pk']
         author = get_object_or_404(User, id=author_id)
         user = request.user
+
+        # Проверка на попытку подписаться на самого себя
+        if user.id == author.id:
+            return Response(
+                {'detail': 'Вы не можете подписаться на самого себя.'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+
+        # Проверка на существующую подписку
         if Subscriptions.objects.filter(user=user, author=author).exists():
             return Response(
                 {'detail': 'Вы уже подписаны на этого пользователя.'},
-                status=status.HTTP_400_BAD_REQUEST)
+                status=status.HTTP_400_BAD_REQUEST
+            )
+
+        # Создание подписки
         subscription = Subscriptions.objects.create(user=user, author=author)
         serializer = self.get_serializer(subscription)
         return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -47,15 +60,22 @@ class SubscribeViewSet(viewsets.ModelViewSet):
         author_id = self.kwargs['pk']
         author = get_object_or_404(User, id=author_id)
         user = request.user
-        subscription = get_object_or_404(
-            Subscriptions,
-            user=user,
-            author=author)
+
+        # Проверка на существование подписки
+        subscription = Subscriptions.objects.filter(user=user,
+                                                    author=author).first()
+        if subscription is None:
+            return Response(
+                {'detail': 'Подписка не найдена.'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+
         subscription.delete()
 
         return Response(
             {'detail': 'Подписка удалена'},
-            status=status.HTTP_204_NO_CONTENT)
+            status=status.HTTP_204_NO_CONTENT
+        )
 
 
 class SubscriptionsViewSet(viewsets.ModelViewSet):
@@ -82,7 +102,6 @@ class SubscriptionsViewSet(viewsets.ModelViewSet):
 
 class RecipeShortLinkView(generics.GenericAPIView):
     hashids = Hashids(min_length=4, salt="SALT")
-
     def get(self, request, id):
         recipe = get_object_or_404(Recipe, id=id)
         short_id = self.hashids.encode(recipe.id)
@@ -112,14 +131,21 @@ class ShoppingCartViewSet(viewsets.ModelViewSet):
 
     def destroy(self, request, *args, **kwargs):
         recipe_id = self.kwargs['pk']
-        user = self.request.user
-        favorite = get_object_or_404(
-            ShoppingCart,
-            recipe__id=recipe_id,
-            user=user)
-        favorite.delete()
+        user = request.user
 
-        return Response({'detail': 'Recipe removed from favorites'},
+        # Проверка на существование записи в корзине
+        shopping_cart_item = ShoppingCart.objects.filter(recipe__id=recipe_id,
+                                                         user=user).first()
+
+        if shopping_cart_item is None:  # Если запись не найдена, возвращаем 404
+            return Response(
+                {'detail': 'Рецепт не найден в корзине.'},
+                status=status.HTTP_404_NOT_FOUND
+            )
+
+        # Удаление записи, если она существует
+        shopping_cart_item.delete()
+        return Response({'detail': 'Рецепт удалён из корзины'},
                         status=status.HTTP_204_NO_CONTENT)
 
 
@@ -153,14 +179,45 @@ class RecipeViewSet(viewsets.ModelViewSet):
     permission_classes = [IsAuthenticatedOrReadOnly]
     queryset = Recipe.objects.all()
     pagination_class = RecipePagination
+    filter_backends = [DjangoFilterBackend]
+    filterset_class = RecipeFilter
+
+    def partial_update(self, request, *args, **kwargs):
+        recipe_id = self.kwargs['pk']
+        recipe = get_object_or_404(Recipe, id=recipe_id)
+
+        # Проверяем, является ли пользователь владельцем рецепта
+        if recipe.author != request.user:
+            return Response(
+                {'detail': 'У вас нет прав на обновление этого рецепта.'},
+                status=status.HTTP_403_FORBIDDEN)
+
+        # Если пользователь владелец, продолжаем с обновлением
+        serializer = self.get_serializer(recipe, data=request.data,
+                                         partial=True)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
     def get_serializer_class(self):
-        if self.request.method == 'GET':
-            return RecipeGetSerializer
-        return RecipePostSerializer
+            if self.request.method == 'GET':
+                return RecipeGetSerializer
+            return RecipePostSerializer
 
     def perform_create(self, serializer):
         serializer.save(author=self.request.user)
+
+    def destroy(self, request, *args, **kwargs):
+        recipe_id = self.kwargs['pk']
+        recipe = get_object_or_404(Recipe, id=recipe_id)
+
+        if recipe.author != request.user:
+            return Response(
+                {'detail': 'У вас нет прав на удаление этого рецепта.'},
+                status=status.HTTP_403_FORBIDDEN)
+        recipe.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class TagViewSet(viewsets.ReadOnlyModelViewSet):
@@ -173,5 +230,6 @@ class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
     pagination_class = None
-    filter_backends = (filters.SearchFilter,)
+    filter_backends = (filters.SearchFilter, DjangoFilterBackend)
+    filterset_class = IngredientFilter
     search_fields = ('^name',)
